Move UserActions to common adapters w/ other User*
