+++ shared/common-adapters/index.desktop.js
@@ -22,5 +22,6 @@ export {default as Terminal} from './terminal'
 export {default as Text} from './text'
 export {default as Usernames} from './usernames'
 export {default as UserBio} from './user-bio'
+export {default as UserActions} from './user-actions'
 export {default as UserCard} from './user-card'
 export {default as UserProofs} from './user-proofs'
