+++ shared/common-adapters/index.native.js
@@ -20,6 +20,7 @@ export {default as TabBar} from './tab-bar'
 export {default as Terminal} from './terminal'
 export {default as Text} from './text'
 export {default as UserBio} from './user-bio'
+export {default as UserActions} from './user-actions'
 export {default as UserCard} from './user-card'
 export {default as UserProofs} from './user-proofs'
 export {default as Usernames} from './usernames'
